http://trac.atomiclabs.com/ticket/60
The OSSP UUID library (http://www.ossp.org/pkg/lib/uuid/) is now a requirement
for pion-platform.

Added createUUID() and createUniqueObjectId() functions to ConfigManager

CodecFactory::addCodec() now auto-generates the Codec's identifier, and 
returns it as the result to the calling function

Added a "name" configuration parameter field for Codecs, which works the same
way as the existing "comment" parameters.