Pion plug-in identifiers are no longer URIs.  Instead, they are now just plain-old UUIDs.  Part of the motivation for this change is to make it easier to refer to these objects using RESTful web services (colons are not allowed in URLs).  Also, each object has a location rooted to the server on which it runs.  Therefore, global reference to each object can be achieved by first referring to the server, followed by a path to the object.  For example:
http://pion-server.example.com:8080/reactors/cc154562-de96-11dc-b285-0016cb926e68

Note that Vocabularies and Vocabulary Terms are NOT changing.  These 
concepts should not be rooted to a particular server, but rather be 
global concepts that can be shared across servers.  In fact, this is 
necessary for the inter-server streaming of events (so that two servers 
can agree and recognize the terms values within Events that are being 
exchanged).  So..  these shall stay URIs (and will probably never be 
UUIDs but rather URLs).