Added placeholder classes for TransformReactor, LogInputReactor, LogOutputReactor, add SQLiteOutputReactor plugins.  Added all of these as well as SQLiteDatabase to the XCode and autoconf build projects. Note that all plugins we plan to include in the 1.0 release at least have placeholders now and are building -- many just don't do anything yet =)
Added sqlite3.h and sqlite3.c files to the repository and the 
SQLiteDatabase project.  Building this "amalgamation" is now the 
recommended way to use the SQLite engine.  It certainly would simplify 
things if this works for all platforms, although it adds a bit of time 
to building the platform.  I figure we can try it out this way and 
switch to using it as an external lib if this doesn't work out.

Removed the MySQLDatabase placeholder files since they will no longer be 
included in our open source distribution and since it was grossly out of
date and not working anyway.