Added a whole bunch of new Reactor web services.  Use "PUT" to change a Reactor's config, "POST" to create a new Reactor and "DELETE" to remove an existing one.  There are also now services to "start" and "stop" Reactors.
Added a bunch of unit tests for the new Reactor services; these are located in 
a new ConfigServiceTests.cpp file.

Added a "Flush" configuration parameter to Codecs, which is used to determine
whether or not the Codec flushes the stream after every write operation.

Updated the LogOutputReactor and FeedService to stop flushing after every write
operation (since this is now handled by Codecs)

Updated the addPlugin() functions so that they only take one parameter now: 
the XML configuration information.  The plugin type is now extracted from the
XML (and must be there!).  This is mainly to remove the redundancy of having the
type both in the function parameter and in the XML config.

You can now start and stop individual Reactors through the ReactionEngine 
using startReactor() and stopReactor(), and they all now have a "Running" 
status, the value of which ("true" or "false") is now returned by the 
"/config/reactors/stats" web service.

The ReactionEngine is now started automatically when the the config file is 
opened.  It should never be stopped, except during shutdown.  Start and stop 
individual reactors instead.

Fixed a bug in ConfigManager::setPluginConfig that was causing a crash if the 
first node in the config list was a "Plugin" node.

Fixed a deadlock bug in the ReactionEngine::removeReactor() function

Misc cleanup and refactoring in unit tests code