http://trac.atomiclabs.com/ticket/123
Increased speed of the LogCodec parser by using the underlying stream 
buffer instead of the istream itself

Changed the clickstream#date vocabulary field to be a shortstring type 
for now, since Boost.DateTime parsing is insanely inefficient.  Looks 
like we will definitely need to write our own date_time parsing routines 
=(

Added a "JustOne" boolean parameter to the LogInputReactor.  If true, it 
reads only the first log entry and then just makes lots of copies of it 
until the reactor is stopped (this is just for testing purposes).

Added an IntAllocTwoThreadTest to test the performace of PionLockedQueue

Added files to XCode project for user authentication in pion-net