http://trac.atomiclabs.com/ticket/123
Added a ReactionScheduler class which is now used for scheduling work in 
the ReactionEngine.  asio's io_service does not perform well under heavy 
load due to an inefficient queue and the triggering of conditions for 
each work event.  This new class fixes these two problems and greatly 
improves performance (on Linux at least) when an event is scheduled to 
be delivered by another thread.

Changed ReactionEngine scheduling so that it now uses multiple threads 
again when there are two or more connections to deliver it to.

Also, fixed a bug in the multi-thread scheduling code that was causing 
it to sometimes post() Events for delivery even when the Reactor was not 
running.