Fixed bug in UserManager that was causing user updates to double-up XML fields in the config.  Now, updates should overwrite any existing fields defined for a user.
Also, if an update is performed, it now checks to see if the password in
the new XML matches the existing encrypted password.  If so, then it knows
that the password should not be changed.  If it doesn't match, it assumes
that the password is plaintext, and it will encrypt and store the encrypted
value.

Added more unit tests to check all the above cases