Updated PionLockedQueue to use a caching freelist when PION_HAVE_LOCKFREE is defined (replaces the ability to use a PionPoolAllocator, which doesn't make as much sense since it is designed more for variable-sized objects)
Added a new variant of PionLockFreeQueue that uses heap memory instead of
fixed memory, and also updated the old version to use boost::lockfree
instead of apr.  Unfortunately, neither of these seem to work when there
are multiple consumers since the assignment of the data during pop() is not
thread-safe.