http://trac.atomiclabs.com/ticket/240
Lots of clean-up on the HTTPParser class to help eliminate redundant code
and make it a lot easier to use.

Updated HTTPMessage::receive() and the HTTPReader, HTTPRequestReader
and HTTPResponseReader classes to use the new HTTPParser -- removing lots of
redundant and confusing code from all of these.

Removed HTTPMessageParser class, as it is no longer necessary