http://trac.atomiclabs.com/ticket/251
Added extra XML config options to platform.xml for the ReactionEngine: 
MaxThreads and MultithreadBranches.

Re-worked some of the Reactor/connection event scheduling code so that it's
now configurable (using MultithreadBranches) instead of hard coded.

Changed the default behavior for Reactor connection branches to NOT use 
multithreading since in most cases this will likely be more efficient.

Fixed a problem with multithreaded connection branching so that it uses the
current thread more often -- there was a bug that caused unnecessary queueing
to other threads to happen when some of the Reactors it was connected to were
not running.