Added a rudimentary EventContainer type
Updated Reactor so that you can deliver an entire EventContainer at once

Updated ConfigManager so that getConfigOption() works for any data types 
supported by boost::lexical_cast (to help eliminate duplicate code)

You can now construct EventFactory objects using an existing Event.  
This will make it use the same EventAllocator as the existing Event.  
Note that it's most efficient when the EventAllocator assigned to the 
current thread is used.  Most likely when a Reactor is generating a new 
Event, it is based on the existing Event it is processing (which most 
likely is using the EventAllocator for the current thread).  So, use 
this whenever you don't have reusable thread-specific EventFactories.