http://trac.atomiclabs.com/ticket/273
Patches to help LogInputReactorTests and ReactorTests pass more often.

The patches in r995 moved the first sleep to occur after the first
event check in getEventsIn() and getEventsOut().  Having the pause AFTER
versus BEFORE the first check seems to aggravate a concurrency bug in 
ReactionEngine that leads to crashing.

Note that this patch doesn't really fix the root of the problem but
just reverts to the old behaviour that avoids it (most of the time).
There is no reason I can see why a pause should be necessary.