http://trac.atomiclabs.com/ticket/647
Added a copy of the latest version of the Boost.Lockfree Library into the
repository, to make it easier to stay up-to-date and in-sync with changes.

Updated PionLockedQueue, PionPoolAllocator, and PionLockFreeQueue to use the
latest version of Boost.Lockfree.

Added documentation for setting up a development environment on
Ubuntu Jaunty (9.04)

Various other documentation updates