DatabaseInserter
Added optional caching of keys, only usable if the table has a "unique" index defined.
Define MaxKeys in the XML (no UI), and DBI will find the unique key, and cache the last n entries.
If a key is already in the cache, it won't be attempted to store in the DB.