pion
Added "-M" flag, to lock-all-memory (mlockall)

This works only on Unix'es -- so it's a dummy flag on Windows.

NOTE: using -M guarantees that all memory is locked in RAM -- which, in turn, guarantees that when/if the system
runs out of memory, Pion will crash (instead of dying slowly, but equally surely).

NOTE: You must run Pion as root (or user with "ulimit -l" set high enough), otherwise you get a nasty error message in the log file.