http://trac.atomiclabs.com/ticket/709
Modified /query/reactors that it returns detailed stats for all reactors.

This required changes in all reactors that implement Reactor::query(), because
the <PionStats> must be sent outside of the call (in ReactionEngine()) instead
of within the Reactor::query() method.

Since some query services don't seem to work properly if they are wrapped
inside of <PionStats>, I also had to add some logic to check to see if any
query parameters are included for /query/reactors/<id>, and to NOT wrap the
response in <PionStats> if there are.

Note that this (slightly) changes the behavior of some query services, which
before may have always wrapped the response inside of <PionStats> even when
parameters were specified