http://trac.atomiclabs.com/ticket/651
PythonReactor checkpoint:

process(), start() and stop() python functions now take a class of type 
Reactor as their first argument.  The new Reactor class has "id" and 
"name" attributes which match the PythonReactor instance's, and a 
deliver() method which can be used to deliver a Python event object to 
the Reactor's downstream connections

All that's left now is the pion event <-> python event conversions!