http://trac.atomiclabs.com/ticket/1136
moved pion_static_lib() and pion_make_static() from with-pion.inc to shared
pion-setup.inc file

I changed the code for the PION_DECLARE_PLUGIN() macro so that it integrates
static-linked plugins directly without requiring a separate list.  Essentially
this eliminates and cleans up a lot of the code and pushes the abstraction one
layer deeper (from PluginManager to PionPlugin) so that all PionPlugin
functions work the same for static or dynamic libraries.

I also added a new PionPluginInstancePtr class which works as a smart pointer
for both the object code for a plugin (static or dynamic) as well as a
heap-allocated instance of the plugin.  This makes it much easier to handle
creation/destruction for code that doesn't use PluginManager.

I also added InterfaceStub.hpp for all plugins built for the unit tests and
added some unit tests for the PionPluginInstancePtr class.